requirements.txt
db-to-sqlite==1.1.1
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
PyJWT==1.7.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.15
sqlite-utils==2.3.1
sqlparse==0.3.1
tabulate==0.8.6
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
python-resize-image